requirements.txt
accelerate==1.7.0
transformers==4.44.0
ftfy==6.1.1
# albumentations==1.3.0
opencv-python==4.8.1.78
einops==0.7.0
pytorch-lightning==1.9.0
tensorboard
scipy==1.11.4
safetensors==0.4.3
omegaconf==2.3.0
altair==4.2.2
easygui==0.98.3
toml==0.10.2
voluptuous==0.13.1
huggingface-hub==0.27
# for Image utils
imagesize==1.4.1
fairscale==0.4.13
invisible-watermark==0.2.0
rich>=13.7.1
diffusers==0.33.1